documentation and cleanup after refactoring to Arduino strategy. Averaging strategy abandoned.
