+++ test.dot
@@ -1,0 +1,37 @@
+
+digraph {
+	layout=dot
+	subgraph cluster_Arduino {
+		color=blue;
+		node [style=filled];
+		"Bunny.ino" ;
+		label = "Ardunio";
+		}
+
+	subgraph cluster_Rpi {
+		color=darkgrey;
+		node [style=filled];
+		"server.py" -> "ardunio_serial.py";
+		label = "Raspberry Pi";
+		}
+
+
+	subgraph cluster_Laptop {
+		color=green;
+		node [style=filled];
+		"Client.py"->"Animate.py", "Stelcom.py";
+		"Stellariutm", "flattend.py";
+		label = "Laptop";
+		}
+
+	BNO055 -> "Bunny.ino"
+	BNO055 [shape=square]
+	"Bunny.ino" -> "ardunio_serial.py" [taillabel = "Serial Print", headlabel = "I2C"]
+	"server.py" ->"Client.py" [taillabel="asyncio", label = "wifi", headlabel = "asyncio" ]
+	"Stelcom.py"-> "Stellariutm"[taillabel = "post", arrowhead = diamond]
+	"ardunio_serial.py"->"server.py"
+	"Stellariutm" [shape=arrow];
+
+
+
+}
